fix #18: Add casts

+++ kernel-4.4/drivers/misc/mediatek/gpu/gpu_rgx/m1.8ED4490469/services/server/devices/rgx/rgxinit.c
@@ -444,7 +444,7 @@ static PVRSRV_ERROR RGXGetGpuUtilStats(PVRSRV_DEVICE_NODE *psDeviceNode,
 #ifndef ENABLE_COMMON_DVFS
 	OSLockAcquire(psDevInfo->hGPUUtilLock);
 #else
-	spin_lock_irqsave(&psDevInfo->sGPUUtilLock, ui64LockFlags);
+	spin_lock_irqsave(&psDevInfo->sGPUUtilLock, (unsigned long)ui64LockFlags);
 #endif
 
 	/* Read the timer before reading the latest stats from the shared
